feat(#314): cloud function to delete slides on deck delete
