+++ cloud/functions/src/index.ts
@@ -4,14 +4,16 @@ import * as admin from 'firebase-admin';
 const app: admin.app.App = admin.initializeApp();
 app.firestore().settings({timestampsInSnapshots: true});
 
-import {applyWatchDeckWrite} from './watch/watch-deck-write';
-import {applyWatchUserDelete} from './watch/watch-user-delete';
+import {applyWatchDeckDelete, applyWatchDeckUpdate} from './watch/watch-deck';
+import {applyWatchUserDelete} from './watch/watch-user';
 
 const runtimeOpts = {
     timeoutSeconds: 120,
     memory: <const> '1GB'
 };
 
-export const watchDeckWrite = functions.runWith(runtimeOpts).firestore.document('decks/{deckId}').onWrite(applyWatchDeckWrite);
+export const watchDeckUpdate = functions.runWith(runtimeOpts).firestore.document('decks/{deckId}').onUpdate(applyWatchDeckUpdate);
 
-export const watchDeleteUser = functions.auth.user().onDelete(applyWatchUserDelete);
+export const watchDeckDelete = functions.firestore.document('decks/{deckId}').onDelete(applyWatchDeckDelete);
+
+export const watchUserDelete = functions.auth.user().onDelete(applyWatchUserDelete);
