+++ cloud/functions/src/watch/delete/delete-decks-slides.ts
@@ -1,12 +1,8 @@
 import * as admin from 'firebase-admin';
 
 import {Deck, DeckData} from '../../model/deck';
-import {Slide} from '../../model/slide';
 
-interface DeckSlides {
-    deckId: string;
-    slides: Slide[] | null;
-}
+import {DeckSlides, deleteSlides, findSlides} from './utils/delete-slides-utils';
 
 export async function deleteDecksSlides(userRecord: admin.auth.UserRecord) {
     if (!userRecord || !userRecord.uid || userRecord.uid === undefined || userRecord.uid === '') {
@@ -70,42 +66,6 @@ function findDecks(userId: string): Promise<Deck[] | null> {
     })
 }
 
-function findSlides(deckId: string): Promise<DeckSlides | null> {
-    return new Promise<DeckSlides | null>(async (resolve, reject) => {
-        try {
-            if (!deckId || deckId === undefined || deckId === '') {
-                resolve();
-                return;
-            }
-
-            const collectionRef: admin.firestore.CollectionReference = admin.firestore().collection(`/decks/${deckId}/slides/`);
-
-            const snapShot: admin.firestore.QuerySnapshot = await collectionRef.get();
-
-            if (snapShot && snapShot.docs && snapShot.docs.length > 0) {
-                const slides: Slide[] = snapShot.docs.map((doc) => {
-                    const id = doc.id;
-                    const ref = doc.ref;
-
-                    return {
-                        id: id,
-                        ref: ref
-                    } as Slide;
-                });
-
-                resolve({
-                    deckId: deckId,
-                    slides: slides
-                });
-            } else {
-                resolve(null);
-            }
-        } catch (err) {
-            reject(err);
-        }
-    })
-}
-
 function findAllSlides(decks: Deck[]): Promise<(DeckSlides | null)[]> {
     return new Promise<(DeckSlides | null)[]>(async (resolve, reject) => {
         try {
@@ -145,50 +105,6 @@ function deleteAllSlides(decksSlides: (DeckSlides | null)[]): Promise<void> {
             if (promises && promises.length > 0) {
                 await Promise.all(promises);
             }
-
-            resolve();
-        } catch (err) {
-            reject(err);
-        }
-    });
-}
-
-function deleteSlides(deckId: string, slides: Slide[] | null): Promise<void> {
-    return new Promise<void>(async (resolve, reject) => {
-        try {
-            if (!slides || slides.length <= 0) {
-                resolve();
-                return;
-            }
-
-            const promises: Promise<void>[] = [];
-            slides.forEach((slide: Slide) => {
-                promises.push(deleteSlide(deckId, slide));
-            });
-
-            if (promises && promises.length > 0) {
-                await Promise.all(promises);
-            }
-
-            resolve();
-        } catch (err) {
-            reject(err);
-        }
-    });
-}
-
-function deleteSlide(deckId: string, slide: Slide): Promise<void> {
-    return new Promise<void>(async (resolve, reject) => {
-        try {
-            if (!slide || !slide.id || slide.id === undefined || slide.id === '') {
-                resolve();
-                return;
-            }
-
-            const collectionRef: admin.firestore.CollectionReference = admin.firestore().collection(`/decks/${deckId}/slides/`);
-            const doc: admin.firestore.DocumentReference = collectionRef.doc(slide.id);
-
-            await doc.delete();
 
             resolve();
         } catch (err) {
