--- cloud/functions/src/watch/watch-deck-write.ts
@@ -1,8 +1,0 @@
-import {Change, EventContext} from 'firebase-functions';
-import {DocumentSnapshot} from 'firebase-functions/lib/providers/firestore';
-
-import {generateDeckScreenshot} from './screenshot/generate-deck-screenshot';
-
-export async function applyWatchDeckWrite(change: Change<DocumentSnapshot>, context: EventContext) {
-    await generateDeckScreenshot(change);
-}
